requirements.txt

scipy~=1.9.0
networkx~=2.8.5
matplotlib==3.5.3
PyYAML~=6.0
numpy<=1.22.4
numba>=0.56.0
imageio>=2.21.1
pytest